Removed version import from setup.py as it was breaking pip install esdc_api (#1)
* Removed version import from setup.py as it was breaking pip install esdc_api 

When package was installed via pip it ended with error missing requests
library, even it is part of the dependencies in setup. Problem was that
import inside __init__.py was loading client module that was importing
requests. This has happened before the dependency was installed.

We have decided to keep the version on two places (setup.py and __init__.py). This way we do not
break things for users and fix the install issue.